Move i3c ROM code into separate file
