+++ rom/src/lib.rs
@@ -14,15 +14,11 @@ Abstract:
 
 #![no_std]
 
-#[cfg(target_arch = "riscv32")]
 mod fuses;
-#[cfg(target_arch = "riscv32")]
 pub use fuses::*;
-
-#[cfg(target_arch = "riscv32")]
-mod riscv;
-#[cfg(target_arch = "riscv32")]
-pub use riscv::*;
+mod rom;
+pub use rom::*;
+mod i3c;
 
 pub trait FatalErrorHandler {
     fn fatal_error(&mut self, code: u32) -> !;
